fix: 修复编译错误和警告
主要修复:
- 修复 Lua 链接错误：添加 setGlobal 等缺失的函数实现
- 重新启用嵌入脚本功能
- 确保 Lua 库正确链接到 runtime 目标

编译警告修复:
- window.cpp: 移除未使用的变量 'len'
- process.cpp: 标记未使用的 'force' 参数
- human.cpp: 修复 char 类型转换警告
- trigger.cpp: 修复有符号/无符号比较警告
- trigger_engine.cpp: 添加 lua_Integer 到 int 的显式转换
- system.cpp: 修复未初始化变量和未使用变量警告
- json.cpp: 移除未使用的异常变量
- kvstore.cpp: 移除未使用的 'now' 变量
- remote_server.cpp: 修复类型转换和未使用参数
- ocr_stub.cpp: 标记未使用的参数
- ml_stub.cpp: 标记未使用的参数
- transport 头文件: 修复未使用参数和类型转换
- emmy_adapter.cpp: 修复 WIN32_LEAN_AND_MEAN 宏重定义警告
- resource_loader.cpp: 使用正确的宽字符转换函数

+++ lib/wingman/src/trigger.cpp
@@ -216,7 +216,7 @@ DWORD WINAPI TriggerManager::checkThread(LPVOID param) {
             if (manager->checkTrigger(trigger)) {
                 // 检查冷却时间
                 DWORD now = GetTickCount();
-                if (now - trigger.lastTriggerTime >= trigger.config.cooldown) {
+                if (now - trigger.lastTriggerTime >= static_cast<DWORD>(trigger.config.cooldown)) {
                     trigger.lastTriggerTime = now;
                     manager->executeActions(trigger.config.actions);
